+++ include/TGUI/Widgets/ListBox.hpp
@@ -360,8 +360,37 @@ namespace tgui
         /// Items that were not given an id simply have an empty string as id.
         ///
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        const std::vector<String>& getItemIds() const;
+        std::vector<String> getItemIds() const;
 
+
+        /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+        /// @brief Store some user data with the item
+        ///
+        /// @param data  User data to store
+        ///
+        /// Examples:
+        /// @code
+        /// listBox->setItemData(idx, "Data"); // Note: type to retrieve with getItemData is 'const char*' here
+        /// listBox->setItemData(idx, 5);
+        /// @endcode
+        /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+        void setItemData(std::size_t index, Any data);
+
+
+        /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+        /// @brief Returns user data stored in the item
+        /// @return Stored data
+        /// @throw std::bad_cast if the template type does not match the type inside the std::any variable passed in setItemData
+        ///        or when the index was too high (which acts as if you access an empty std::any variable).
+        /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+        template <typename T>
+        T getItemData(std::size_t index) const
+        {
+            if (index < m_items.size())
+                return AnyCast<T>(m_items[index].data);
+            else
+                throw std::bad_cast();
+        }
 
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
         /// @brief Changes the height of the items in the list box
@@ -615,9 +644,14 @@ namespace tgui
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     protected:
 
-        // This contains the different items in the list box
-        std::vector<Text>       m_items;
-        std::vector<String> m_itemIds;
+        struct Item
+        {
+            Text text;
+            Any data;
+            String id;
+        };
+
+        std::vector<Item> m_items;
 
         // What is the index of the selected item?
         // This is also used by combo box, so it can't just be changed to a pointer!